requirements.txt
alive-progress==1.6.1
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-extensions==3.0.9
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
